fix(config): allow fractional handoff_window_hours (e.g. 0.05 for smoke testing)

+++ src/core/salon-config-schema.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 export const SalonConfigSchema = z.object({
   response_delay_ms: z.number().int().positive().default(40_000),
   llm_model: z.string().optional(),
-  handoff_window_hours: z.number().int().positive().default(4),
+  handoff_window_hours: z.number().positive().default(4),
   booking_link_dedup_window: z.number().int().positive().default(3),
   max_words_per_message: z.number().int().positive().default(40),
   max_emojis: z.number().int().nonnegative().default(2),
